use_case_1/requirements.txt
black==22.10.0
django==4.1.4
djangorestframework==3.14.0
flake8==6.0.0
pytest==7.2.0
requests==2.28.1
requests-mock==1.10.0
